fix: create script directory for python file
allows for automatic1111 to be able to grab it via git url
